Bug 8715: Receive all serials for a year.
This patch adds
- 2 links in the table header (select all and clear all).
- Datatable on these tables

Test plan:
Try to select all serials for a year and receive them.
Test there is no regression (ergonomic) on this page

Signed-off-by: Owen Leonard <oleonard@myacpl.org>

Bug 8715 [Follow-up] Receiving all serials for a year

Follow-up: For consistency, I would prefer to have the
column of checkboxes at the beginning of the table and the
select/clear links above.

Signed-off-by: Jonathan Druart <jonathan.druart@biblibre.com>

Bug 8715: Force the default sort order to desc

Signed-off-by: Paul Poulain <paul.poulain@biblibre.com>
Signed-off-by: Jared Camins-Esakov <jcamins@cpbibliography.com>