Revert "Bug 8089: Use Koha::Cache for all caching"
This reverts commit 215abc8024d93f1b6a10f26427a65bd64d106e52.

The 3 patches for bug 8089 have been reverted, because they break
jenkins & Koha.
A follow-up has been provided, but it does not solve the problem on my
test server, it just changes the error message.

After a discussion with jared, Dobrica should work on another patch, so
the best option is to revert.