Bug 8840 - [SIGNED-OFF] ubuntu-pkg-check.sh fix and extend functionality
First, reworked it into a newer script with parameters:
   ubuntu-packages.sh -r
This will generate output identical to the old script, except
that it is fixed for non-english setups.
   ubuntu-packages.sh -h
This adds a brief help, which all scripts should have.
   ubuntu-packages.sh -ic
This helps handle a multi-arch problem by printing out a
relevant command for installing missing dependencies.

Signed-off-by: Jared Camins-Esakov <jcamins@cpbibliography.com>
Signed-off-by: Magnus Enger <magnus@enger.priv.no>
Works as advertised, see Bugzilla for details.
Signed-off-by: Jared Camins-Esakov <jcamins@cpbibliography.com>