Bug 8923 [Revised] Add ids to repeated pagination menus to allow for CSS customization
This patch adds ids to each pagination menu on the OPAC search
results page so that they can be customized via CSS or JavaScript.

To test, modify OpacUserCSS to hide ids "top-pages" and "bottom-pages"
and confirm that they are hidden.

Also fixed is an unescaped ampersand introduced by Bug 8209.

Revision removes debugging changes to an unrelated file.

Signed-off-by: Nicole C. Engard <nengard@bywatersolutions.com>
Signed-off-by: Paul Poulain <paul.poulain@biblibre.com>