Bug 5521 - item edit form appears if no items found
This patch changes the template so that if no items are
found the page doesn't show "The following barcodes were found"
and doesn't show the item edit form.

To test, submit a batch which contains no valid barcodes.
Compare to a batch which contains some or all valid barcodes.

Signed-off-by: Nicole C. Engard <nengard@bywatersolutions.com>
All tests succeed.

Signed-off-by: Paul Poulain <paul.poulain@biblibre.com>
Signed-off-by: Jared Camins-Esakov <jcamins@cpbibliography.com>