Bug 9003 - Add css fix for opac items with strong and em
Currently, if you have nested strong and em tags in the opac, you will only get the properties of the inner tag. This seems to be due to reset-fonts-grids.css resetting the font weight and style (for both strong and em) to normal.

So if you have strong as your outer tag and em as your inner tag, you'll have italics but no bold.

To test:

Before applying the patch...

1) Put <em> and <strong> tags around some text in the opacnav, opac news, opacheader, or any other area of the opac where you can enter html text.
2) Look at the opac and notice that only the inner tag styles are being applied to the text

Apply the patch

3) Refresh the opac (you may need to bypass or clear your browser cache for the new style to pick up)
4) Note that both the inner and outer <em> and <strong> tags are being applied to the text (i.e. your text should be bold and italicized)

Signed-off-by: Bernardo Gonzalez Kriegel <bgkriegel@gmail.com>
Small change that works. No errors.

Signed-off-by: Marcel de Rooy <m.de.rooy@rijksmuseum.nl>
Signed-off-by: Jared Camins-Esakov <jcamins@cpbibliography.com>