Bug 8955 [Follow-up] Self checkout should degrade gracefully w/o JavaScript
This follow-up fixes some instance of unnecessary JavaScript
which I missed the first time around: The renew and check-in
buttons on the list of checkouts. The functionality can be
easily accomplished without JS.

Also corrected:

- Capitalization error
- Miss-named variable causing the "too many checkouts" not
  to be displayed
- Misplaced [% END %] tags hidding "Return to account" button
  on some dialogs.
- Corrected variable scope error which prevented the "return"
  button from appearing.

To test, log into self checkout using an account which has
checkouts. Testing renewing items both by scanning barcodes
and by clicking the "renew" button in the list of checkouts.

Renew something until it reaches its limit and test checking
it out by scanning the barcode. Test the "check in" button.

Signed-off-by: Chris Cormack <chris@bigballofwax.co.nz>
Signed-off-by: Paul Poulain <paul.poulain@biblibre.com>
Signed-off-by: Jared Camins-Esakov <jcamins@cpbibliography.com>