Bug 8893 [Follow-up] show more information on printout
This follow-up removes the old link to detailprint.pl and
replaces it with an empty container as a hook for the YUI
button function.

The appearance and functionality should not be any
different.

Signed-off-by: Melia Meggs <melia@bywatersolutions.com>
Signed-off-by: Jonathan Druart <jonathan.druart@biblibre.com>
This patch removes a ref to detailprint.pl.
With JS a "Print" button is displayed (else it is not).
Signed-off-by: Jared Camins-Esakov <jcamins@cpbibliography.com>