Bug 9130 follow up: add statesave cookie and "all" option for parcel drop-down menu
This patch sets the bStateSave parameter for the "Pending orders" and
"Already received" Datatables in Acquisitions, which uses a cookie to
save the "pagination information, display length, filtering and sorting".

It also sets the iCookieDuration parameter to 1000 days (the default
value is 2 hours otherwise).

It also adds an "all" option to the dro-down menu using the aLengthMenu
parameter.

Signed-off-by: Kyle M Hall <kyle@bywatersolutions.com>
Signed-off-by: Jonathan Druart <jonathan.druart@biblibre.com>
Signed-off-by: Jared Camins-Esakov <jcamins@cpbibliography.com>