Bug 9267 - Ubuntu 12.04 install docs broken
Current install instructions are broken in the (1.6) step. On amd64
architectures it breaks the whole dependency tree, pulling i386 packages.

This patch replaces the 'dpkg --set-selections' combo and replacess some spaces
for tabs in one line.

This should be cherry-picked for other releases that have 12.04 install instructions
as the problem is related to multiarch.

Regards
To+

Signed-off-by: Galen Charlton <gmc@esilibrary.com>
Signed-off-by: Elliott Davis <elliott@bywatersolions.com>
Signed-off-by: Jared Camins-Esakov <jcamins@cpbibliography.com>