Bug 8557 - Unexpected behavior with scanners
This is an alternative to the original implementation - this one works in chrome as well as firefox.

To test:

1)
Set syspref 'AcqCreateItem' to 'Create Item when receiving an order.'

2)
Create a basket for a vendor, add an order line.

3)
Search for this vendor using Vendor search

4)
Receive Shipment for this vendor and choose the title you ordered in 2)

5) You will get the dialog to create related item(s)

6)
Fill in Item 0 through o (see screenshot)

7)
Scan Barcode for field p - Barcode

7)
Result: Scanner sends a [return], form closes and you had no chance to fill in fields t - z

Signed-off-by: Marc Veron <veron@veron.ch>

Tested with Chrome Version 23.0.1271.97 m  and Firefox 17.0.1, both behave as expected.

Signed-off-by: Katrin Fischer <Katrin.Fischer.83@web.de>
Tested with Firefox and Chromium in Ubuntu.

Additional test done:
1) Add a new subscription, choose 'receive adds items'
2) Receive an issue, fill out $p with a barcode
3) Make sure hitting enter in the barcode field does not send the form.

All tests pass and QA script pass.
Signed-off-by: Jared Camins-Esakov <jcamins@cpbibliography.com>