Bug 9283 - Change structure of export checkouts form
This patch changes the structure of the export checkouts form
so that it is a little more linear. This change moves generation of the
export options from JavaScript to the markup, eliminating an instance
of dependence on YUI menus.

To test, enable checkout exports by specifying a value for
ExportWithCsvProfile or  ExportRemoveFields. Load a patron with
checkouts in circulation. Try the various checkout export options. Each
should function correctly.

Signed-off-by: Jonathan Druart <jonathan.druart@biblibre.com>
Signed-off-by: Marcel de Rooy <m.de.rooy@rijksmuseum.nl>
Signed-off-by: Jared Camins-Esakov <jcamins@cpbibliography.com>