Bug 9323 [Counter-proposal] Untranslatable string in serials search
The translator seems to have a problem with the _() function inside
HTML. Rather than move only the translatable string to the <script>
block, why not move all of the JavaScript to the <script> tag? This
keeps things a little more organized and lets us follow the guideline to
always use double quotes with _().

Same test plan as Jonathan's patch:

- translate update LANG
- Check in your po file that the string "Are you sure you want to
  reopen this subscription?" does not exist.
- Apply the patch
- translate update LANG
- Check in your po file that the string is translatable

Signed-off-by: Jonathan Druart <jonathan.druart@biblibre.com>
Signed-off-by: Katrin Fischer <Katrin.Fischer.83@web.de>
String is now translatable.
Signed-off-by: Jared Camins-Esakov <jcamins@cpbibliography.com>