Bug 9215 - non translatable string in manage-marc-import
Status information about staged MARC record batches was being pulled
directly from the import_batches table where they are stored as English
strings. This patch puts a check on the status value into the template
so that translatable strings can be embedded.

To test, apply the patch and view both the table of staged MARC record
batches and details about individual batches. In the default 'en'
translation you should see import statuses displayed with a capital
letter ("Staged"). This indicates that the status is now being pulled
from the template.

Signed-off-by: Chris Cormack <chrisc@catalyst.net.nz>
Signed-off-by: Jonathan Druart <jonathan.druart@biblibre.com>
Signed-off-by: Jared Camins-Esakov <jcamins@cpbibliography.com>