Bug 9366: add widget to bib editor to control display of the MARC tags
If the syspref hide_marc is set to "don't display", tags were not shown
by default.

This patch adds a checkbox to hide/show tags directly on the
cataloguing page. The value is stored in a cookie.

Test plan:
Enable/disable the sysprefs hide_marc and advancedMARCeditor in order to
check if the display is according with what the user want.
Reload the page and verify the value is kept.
The first value of the cookie is retrieved from the hide_marc pref.

Bonus:
This patch removes the link "Show MARCtag documentation links" if the
syspref advancedMARCeditor is enabled.

Signed-off-by: Joel Aloi <aloi54@live.fr>
Signed-off-by: Katrin Fischer <Katrin.Fischer.83@web.de>
This works as described, the editor will remember the choices
made by the cataloguer.

Passes all tests and QA script.

Notes:
- The error messages for mandatory fields upon saving only
give the tag number and not the description, so this might require
to turn back on the tag display.

- The advancedMARCeditor doesn't display the documentation links before
and after the patch was applied. I think it could still be useful for
an advanced cataloger to be able to look up information on a field quickly.

Signed-off-by: Galen Charlton <gmc@esilibrary.com>