Bug 9062 Fix format for discount rate for vendors
On the vendor display and edit screens, the tax rate information was
formatted to 1 decimal point, while the discount rate was not.

This patch adds a format string to the discount information so that
it is formatted in the same way.

Signed-off-by: Hugh Davenport <hugh@davenport.net.nz>
Signed-off-by: Chris Cormack <chris@bigballofwax.co.nz>
Signed-off-by: Katrin Fischer <Katrin.Fischer.83@web.de>
Makes display on vendor detail page more consistent and
passes all tests.
Signed-off-by: Jared Camins-Esakov <jcamins@cpbibliography.com>