Bug 9450 [Revised] Use DataTables in cataloging Z39.50 search results
Replace the tablesorter plugin with the DataTables plugin in cataloging
Z39.50 search results.

Structural changes were made to the table markup to deal with a
DataTables bug which prevents it from properly parsing a <th> with a
colspan.

To test, go to cataloging and perform a Z39.50 search. Confirm that
table sorting works correctly. In this configuration there
should be no pager and no results filtering.

Revision corrects template path to datatables assets.

Signed-off-by: Bernardo Gonzalez Kriegel <bgkriegel@gmail.com>

Comment: Sorting works. No errors.
Signed-off-by: Jonathan Druart <jonathan.druart@biblibre.com>
Signed-off-by: Jared Camins-Esakov <jcamins@cpbibliography.com>